added permissions to api/views/boards.py

+++ api/views/board.py
@@ -1,10 +1,9 @@
 from rest_framework import status
 from rest_framework.response import Response
-from rest_framework.viewsets import ViewSet
 from rest_framework.views import APIView
 from tasks.models import BoardModel, WorkSpaceModel
 from ..serializers.board import BoardSerializer
-from rest_framework.permissions import IsAuthenticatedOrReadOnly, IsAuthenticated
+from permissions import IsWorkSpaceOwner, IsBoardOwner, IsWorkSpaceMember
 
 # class BoardViewSet(ViewSet):
 #     def list(self, request):
@@ -28,19 +27,23 @@ from rest_framework.permissions import IsAuthenticatedOrReadOnly, IsAuthenticate
 
 
 class BoardListView(APIView):
-    # permission_classes = (IsAuthenticated,)
+    permission_classes = [IsWorkSpaceMember]
 
     def get(self, request, workspace_id):
         workspace = WorkSpaceModel.objects.get(pk=workspace_id)
+        self.check_object_permissions(self, workspace)
         all_board = BoardModel.objects.filter(workspace=workspace)
         ser_data = BoardSerializer(instance=all_board, many=True)
         return Response(data=ser_data.data, status=status.HTTP_200_OK)
 
 
 class BoardCreateView(APIView):
+    permission_classes = [IsWorkSpaceOwner | IsWorkSpaceMember]
+
     def post(self, request, workspace_id):
         data = BoardSerializer(data=request.POST)
         workspace = WorkSpaceModel.objects.get(pk=workspace_id)
+        self.check_object_permissions(self, workspace)
         if data.is_valid():
             data.save(commit=False)
             data.workspce = workspace
@@ -50,8 +53,12 @@ class BoardCreateView(APIView):
 
 
 class BoardUpdateView(APIView):
+    permission_classes = [IsWorkSpaceOwner | IsWorkSpaceMember | IsBoardOwner]
+
     def put(self, request, pk):
         board = BoardModel.objects.get(pk=pk)
+        self.check_object_permissions(self, board.workspace)
+        self.check_object_permissions(self, board)
         ser_data = BoardSerializer(instance=board, data=request.POST, partial=True)
         if ser_data.is_valid():
             ser_data.save()
@@ -60,7 +67,11 @@ class BoardUpdateView(APIView):
 
 
 class BoardDeleteView(APIView):
+    permission_classes = [IsWorkSpaceOwner | IsWorkSpaceMember | IsBoardOwner]
+
     def delete(self, request, pk):
         board = BoardModel.objects.get(pk=pk)
+        self.check_object_permissions(self, board.workspace)
+        self.check_object_permissions(self, board)
         board.delete()
         return Response({"message": 'you successfully delete board'}, status=status.HTTP_200_OK)
